updating names
