+++ BillingSystem/logs/templates/logs/print_details.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <title>Print Details</title>
+    </head>
+    <body>
+       
+        <h2>Fill the details to be printed</h2>
+        <form method="post">
+            {% csrf_token %}
+            <table>
+                {{ form }}
+            </table>
+            {{form.non_field_errors}}
+            <button type="submit">Submit</button>
+        </form>
+    </body>
+</html>
